feat: 新增批量业务VO的增改DTO并引入excel依赖
1. 为近40个业务视图对象新增LocalDateTime导入
2. 新增对应业务的Post和Put请求DTO类
3. 引入ydsz-common-excel公共依赖
4. 修复搜索排序的类型强制转换问题

+++ ydsz-backend/ydsz-common/ydsz-common-search/src/main/java/com/njydsz/common/search/service/BusinessRanker.java
@@ -46,7 +46,7 @@ public class BusinessRanker {
 
         String keyword = request.getKeyword() != null ? request.getKeyword().trim().toLowerCase() : "";
 
-        hits.forEach(hit -> hit.setScore(calculateScore(hit, keyword)));
+        hits.forEach(hit -> hit.setScore((float) calculateScore(hit, keyword)));
 
         hits.sort(Comparator.comparingDouble(SearchHit::getScore).reversed());
         return hits;
